\name{skewt.lines}
\alias{skewt.lines}
\title{Overlays data on a SKEW-T, log p axis}
\description{Overlays observations as lines on a SKEW-T, log p axis
             (as created by \code{skewt.axis}). }
\usage{ skewt.lines(temp, pressure, ...) }
\arguments{
  \item{temp}{ Temperature in degrees C. }
  \item{pressure}{ Pressure in millibars }
  \item{...}{ Any graphical arguments }
}
\details{\code{skewt.lines} overlays observations on a SKEW-T, log p axis}

\seealso{ 
\code{\link{skewtPlot}}
          }



\keyword{ aplot }
