\name{RadioSonde internal }

\alias{satlft}
\alias{skewt.data}
\alias{skewtx}
\alias{skewty}
\alias{wobf}

\title{
  RadioSonde internal and secondary functions
}
\description{
Listed below are supporting functions for the major methods in RadioSonde.
}
\usage{

satlft(thw, p)

skewt.data 

skewtx(temp, ycoord)

skewty(pres) 

wobf

}

\keyword{internal}
\keyword{misc}
