\name{RadioSonde internal }

\alias{satlft}
\alias{skewt.data}
\alias{skewtx}
\alias{skewty}
\alias{sonde.data}
\alias{Tdew2q}
\alias{wobf}

\title{
  RadioSonde internal and secondary functions
}
\description{
Listed below are supporting functions for the major methods in RadioSonde.
}
\usage{

satlft(thw, p)

skewt.data 

skewtx(temp, ycoord)

skewty(pres) 

sonde.data

Tdew2q(temp, td, p)

wobf

}

\keyword{internal}
\keyword{misc}
